[1주차] 객체지향 코드 연습 (fuirian)
