+++ docs/README.md
@@ -1,5 +1,5 @@
-1. 당첨번호입력
-2. 보너스번호입력
-3. 금액입력
-4. 금액에 따른 랜덤 로또 번호 출력
-5. 결과 출력
+1. 당첨번호1~45의 숫자 6개 입력하기
+2. 당첨번호와 중복되지 않는 보너스번호 1~45의 숫자  1개 입력하기
+3. 1000원 단위의 금액 입력하기
+4. 입력된 금액/1000 개의 랜덤 로또 번호 출력하기
+5. 랜덤 로또 번호와 당첨번호, 보너스 번호를 매치하여 결과 출력하기
